configuring auth-service, gateway, basic thymeleaf page done

+++ stock-service/src/main/java/com/personal/project/stockservice/model/dto/response/CompleteStockDTO.java
@@ -7,18 +7,32 @@ import java.math.BigDecimal;
 @Data
 public class CompleteStockDTO {
 	private String stockId;
+
 	private String stockName;
+
 	private Long date;
+
 	private BigDecimal todayClosingPrice;
+
 	private BigDecimal yesterdayClosingPrice;
+
 	private BigDecimal priceGap;
+
 	private BigDecimal priceGapPercent;
-	private BigDecimal todayTradingVolumePiece;
+
+	private Long todayTradingVolumePiece;
+
 	private BigDecimal todayTradingVolumeMoney;
+
 	private BigDecimal upperShadow;
+
 	private BigDecimal realBody;
+
 	private BigDecimal lowerShadow;
+
 	private BigDecimal ma5;
+
 	private BigDecimal ma10;
+
 	private BigDecimal ma20;
 }
